requirements.txt
# Python requirements for ZMQ audio exfiltration testing
pyzmq>=25.0.0
numpy>=1.20.0  # Optional, for better audio generation in test producer
